Clear the gid flags before processing them
Also read the gids as unsigned integers since that's how Tiled writes them
to prevent the number going negative when the highest flag is used.

+++ src/utils/xml.cpp
@@ -97,7 +97,7 @@ namespace XML
         xmlChar *prop = xmlGetProp(node, BAD_CAST name);
         if (prop)
         {
-            ret = atoi((char*) prop);
+            ret = atol((char*) prop);
             xmlFree(prop);
         }
